+++ example/lib/main.dart
@@ -4,6 +4,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:nonso/nonso.dart' as nonso;
+import 'package:nonso/nonso.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -55,8 +56,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
   @override
   Widget build(BuildContext context) {
-    // Use FirebaseAuth.instance.currentUser! to get information about the signed in user
-    final userUID = FirebaseAuth.instance.currentUser!.uid;
+    final User currentSignedInUser = context.read<AuthBloc>().state.user!;
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Theme.of(context).colorScheme.inversePrimary,
@@ -67,7 +67,7 @@ class _MyHomePageState extends State<MyHomePage> {
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
             const Text('You have pushed the button this many times:'),
-            Text('User UID: $userUID'),
+            Text('User UID: ${currentSignedInUser.uid}'),
             Text(
               '$_counter',
               style: Theme.of(context).textTheme.headlineMedium,